Refactor ResolvableTimeRange into a util class
